+++ src/test/php/predicate/PredicateTest.php
@@ -109,16 +109,6 @@ class PredicateTest extends \PHPUnit_Framework_TestCase
 
     /**
      * @test
-     * @deprecated  since 1.4.0, will be removed with 2.0.0
-     */
-    public function negateReturnsNegatePredicate()
-    {
-        $predicate = new FooPredicate();
-        assert($predicate->negate(), isInstanceOf(NegatePredicate::class));
-    }
-
-    /**
-     * @test
      * @since  1.4.0
      */
     public function everyPredicateCanBeNegated()